requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic==2.10.3
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
authlib==1.3.0
requests==2.31.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
alembic==1.14.0
boto3==1.35.73
